+++ packages/app-desktop/src/lib/themes/darkwrite-default.ts
@@ -20,4 +20,5 @@ export const DarkwriteDefault: Theme = {
     border: "0 0% 26%",
     focusRing: "212.7 26.8% 50%",
     mode: "dark",
+    star: "50 97% 63%",
 };