+++ packages/app-desktop/src/lib/themes/index.ts
@@ -1,5 +1,17 @@
-import { CatppuccinMocha } from "./catppuccin";
+import {
+    CatppuccinMocha,
+    CatppuccinLatte,
+    CatppuccinMacchiato,
+    CatppuccinFrappe,
+} from "./catppuccin";
 import { DarkwriteDim } from "./darkwrite-black";
 import { DarkwriteDefault } from "./darkwrite-default";
 
-export const DEFAULT_THEMES = [DarkwriteDefault, CatppuccinMocha, DarkwriteDim];
+export const DEFAULT_THEMES = [
+    DarkwriteDefault,
+    DarkwriteDim,
+    CatppuccinMocha,
+    CatppuccinMacchiato,
+    CatppuccinFrappe,
+    CatppuccinLatte,
+];